+++ resources/assets/js/components/enso/bulma/Tab.vue
@@ -26,6 +26,10 @@ export default {
             type: Boolean,
             default: false,
         },
+        disabled: {
+            type: Boolean,
+            default: false,
+        },
     },
 
     computed: {